+++ projects/hx-ui/src/lib/tabular/tabular.component.ts
@@ -115,7 +115,7 @@ export class TabularComponent implements OnInit, DoCheck {
   protected _callback: Function;
   protected _config: ITabularConfig;
   protected _searchTerm: string;
-  private _isSorting = false;
+  private _isMutatingInternally = false;
   private _initialLoad = true;
 
 
@@ -134,12 +134,12 @@ export class TabularComponent implements OnInit, DoCheck {
     if (!_.isEqual(this.rows, this.oldRows)) {
       this.oldRows = _.cloneDeep(this.rows);
 
-      // on sorting we dont need change detection to run and call sorting again
-      if (!this._isSorting ) {
+      // on sorting or checkbox toggling we don't need change detection to run and call sorting again
+      if (!this._isMutatingInternally ) {
         this.orderByData(false);
       }
 
-      this._isSorting = false;
+      this._isMutatingInternally = false;
 
       this.checkSelectAllState();
     }
@@ -163,10 +163,11 @@ export class TabularComponent implements OnInit, DoCheck {
   }
 
   toggleSelectAll = ($event) => {
-    for (let i = 0; i < this.rows.length; i++) {
-      this.rows[i].checked = this.selectAll;
-    }
-    this.onCheckAll.emit(this.selectAll);
+      for (let i = 0; i < this.rows.length; i++) {
+        this._isMutatingInternally = true;
+        this.rows[i].checked = this.selectAll;
+      }
+      this.onCheckAll.emit(this.selectAll);
   }
 
 
@@ -184,6 +185,7 @@ export class TabularComponent implements OnInit, DoCheck {
     }
 
     const oldSelectAll = this.selectAll;
+    this._isMutatingInternally = true;
     this.selectAll =  (this.rows.length === count);
     if (oldSelectAll !== this.selectAll) {
       this.onCheckAll.emit(this.selectAll)
@@ -274,7 +276,7 @@ export class TabularComponent implements OnInit, DoCheck {
   private orderByData(emitSortEvent: boolean) {
     if (this.config.sortBy.length > 0) {
       if (!this.config.remoteSorting) {
-        this._isSorting = true;
+        this._isMutatingInternally = true;
         this.rows = [...this.rows]; // Required as array-sort-by mutates the original array
         this.sortByService.sortBy(this.rows, this.config.sortBy);
       }
@@ -345,8 +347,9 @@ export class TabularComponent implements OnInit, DoCheck {
     return (action.children && action.children.length > 0);
   }
 
-  trackByFn(index, action) {
-    return index; // or action.id
+  trackByFn(index, item) {
+    return (item.id) ? item.id : index;
   }
+
 
 }