+++ src/app/core/filters/filters.code.ts
@@ -190,7 +190,8 @@ export class FiltersCode {
         id: 'searchFilter',
         type: FilterType.Search,
         label: 'Filter by name',
-        callback: [this.onSearchFilterHandler]
+        callback: [this.onSearchFilterHandler],
+        width: this.getSearchWidth('Filter by name')
       }
     ];
     onFilterChangeEvent$ = new Subscription();
@@ -215,7 +216,26 @@ export class FiltersCode {
     toggleCollapsed() {
       this.collapsed = !this.collapsed;
     }
+
+    // dynamically calculate an appropriate width (rem)
+    getSearchWidth(label: string) {
+      const min = 8;
+      const max = 12;
   
+      // based on root html font size
+      const charwidth = .42;
+  
+      // base search filter
+      const filter = 3.9;
+  
+      const calc = label.length * charwidth + filter;
+  
+      if(min <= calc && calc <= max) {
+        return calc;
+      }
+  
+      return min > calc ? min : max;
+    }
   }